requirements.txt
requests>=2.31.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
umap-learn>=0.5.6
plotly>=5.20.0
